only delete timers on workflow deletion
