+++ common/dynamicconfig/dynamicproperties/constants.go
@@ -1718,16 +1718,15 @@ const (
 	// Default value: true
 	// Allowed filters: N/A
 	EnableGRPCOutbound
-	// EnableTimerCleanupOnWorkflowClose enables cleanup of workflow timer tasks when a workflow
-	// closes before its timers fire. When enabled, tracked timer tasks are deleted on a
-	// best-effort basis at workflow close time (async, errors are logged but do not block
-	// the close), and again at retention-based deletion as a safety net. This is a feature
-	// flag intended to be defaulted to true once validated.
-	// KeyName: system.enableTimerCleanupOnWorkflowClose
+	// EnableWorkflowTimerTaskCleanup enables deletion of tracked workflow timer tasks when
+	// the workflow execution record is cleaned up at the end of the retention period. This
+	// prevents orphaned timer task rows from accumulating in Cassandra. Feature flag,
+	// intended to be defaulted to true once validated.
+	// KeyName: system.enableWorkflowTimerTaskCleanup
 	// Value type: Bool
 	// Default value: false
 	// Allowed filters: N/A
-	EnableTimerCleanupOnWorkflowClose
+	EnableWorkflowTimerTaskCleanup
 	// EnableSQLAsyncTransaction is the key for enabling async transaction
 	// KeyName: system.enableSQLAsyncTransaction
 	// Value type: Bool
@@ -2922,14 +2921,14 @@ const (
 	// Default value: 5m (5*time.Minute)
 	// Allowed filters: N/A
 	StandbyClusterDelay
-	// TimerDeletionOnWorkflowCloseMinTTL is the minimum remaining time before a timer task
-	// is worth explicitly deleting. Timers scheduled to fire within this window are skipped —
-	// they will fire and clean up naturally.
-	// KeyName: history.timerDeletionOnWorkflowCloseMinTTL
+	// WorkflowTimerTaskCleanupMinTTL is the minimum remaining time before a timer task is
+	// worth explicitly deleting at workflow execution cleanup. Timers scheduled to fire within
+	// this window are skipped — they will fire and clean up naturally.
+	// KeyName: history.workflowTimerTaskCleanupMinTTL
 	// Value type: Duration
 	// Default value: 240h (10 days)
 	// Allowed filters: N/A
-	TimerDeletionOnWorkflowCloseMinTTL
+	WorkflowTimerTaskCleanupMinTTL
 	// StandbyTaskMissingEventsResendDelay is the amount of time standby cluster's will wait (if events are missing)before calling remote for missing events
 	// KeyName: history.standbyTaskMissingEventsResendDelay
 	// Value type: Duration
@@ -4606,9 +4605,9 @@ var BoolKeys = map[BoolKey]DynamicBool{
 		Description:  "EnableGRPCOutbound is the key for enabling outbound GRPC traffic",
 		DefaultValue: true,
 	},
-	EnableTimerCleanupOnWorkflowClose: {
-		KeyName:      "system.enableTimerCleanupOnWorkflowClose",
-		Description:  "Enables cleanup of workflow timer tasks when a workflow closes before its timers fire. Deletion at close time is best-effort (async, errors logged but do not block the close); retention-based deletion serves as a safety net. Feature flag, intended to be defaulted to true once validated.",
+	EnableWorkflowTimerTaskCleanup: {
+		KeyName:      "system.enableWorkflowTimerTaskCleanup",
+		Description:  "Enables deletion of tracked workflow timer tasks when the execution record is cleaned up at the end of the retention period. Prevents orphaned timer rows from accumulating in Cassandra. Feature flag, intended to be defaulted to true once validated.",
 		DefaultValue: false,
 	},
 	EnableSQLAsyncTransaction: {
@@ -5617,9 +5616,9 @@ var DurationKeys = map[DurationKey]DynamicDuration{
 		Description:  "StandbyClusterDelay is the artificial delay added to standby cluster's view of active cluster's time",
 		DefaultValue: time.Minute * 5,
 	},
-	TimerDeletionOnWorkflowCloseMinTTL: {
-		KeyName:      "history.timerDeletionOnWorkflowCloseMinTTL",
-		Description:  "Minimum remaining time before a timer task is worth explicitly deleting on workflow close. Timers firing within this window are skipped and will clean up naturally.",
+	WorkflowTimerTaskCleanupMinTTL: {
+		KeyName:      "history.workflowTimerTaskCleanupMinTTL",
+		Description:  "Minimum remaining time before a timer task is worth explicitly deleting at workflow execution cleanup. Timers firing within this window are skipped and will clean up naturally.",
 		DefaultValue: time.Hour * 24 * 10,
 	},
 	StandbyTaskMissingEventsResendDelay: {
