+++ schema/cassandra/cadence/versioned/v0.47/manifest.json
@@ -1,7 +1,7 @@
 {
   "CurrVersion": "0.47",
   "MinCompatibleVersion": "0.47",
-  "Description": "Adding workflow_timer_tasks blob fields to track workflow-level timer tasks for cleanup during workflow deletion",
+  "Description": "Adding workflow_timer_tasks blob fields to track timer tasks associated with a workflow execution for cleanup when the workflow closes before they fire",
   "SchemaUpdateCqlFiles": [
     "workflow_timer_task_info.cql"
   ]
