+++ service/history/config/config_test.go
@@ -75,7 +75,7 @@ func TestNewConfig(t *testing.T) {
 		"ShutdownDrainDuration":                                {dynamicproperties.HistoryShutdownDrainDuration, time.Second},
 		"WorkflowDeletionJitterRange":                          {dynamicproperties.WorkflowDeletionJitterRange, 20},
 		"DeleteHistoryEventContextTimeout":                     {dynamicproperties.DeleteHistoryEventContextTimeout, 21},
-		"EnableTimerCleanupOnWorkflowClose":                    {dynamicproperties.EnableTimerCleanupOnWorkflowClose, true},
+		"EnableWorkflowTimerTaskCleanup":                       {dynamicproperties.EnableWorkflowTimerTaskCleanup, true},
 		"MaxResponseSize":                                      {nil, maxMessageSize},
 		"HistoryCacheInitialSize":                              {dynamicproperties.HistoryCacheInitialSize, 22},
 		"HistoryCacheMaxSize":                                  {dynamicproperties.HistoryCacheMaxSize, 23},
@@ -95,7 +95,7 @@ func TestNewConfig(t *testing.T) {
 		"StandbyClusterDelay":                                  {dynamicproperties.StandbyClusterDelay, time.Second},
 		"StandbyTaskMissingEventsResendDelay":                  {dynamicproperties.StandbyTaskMissingEventsResendDelay, time.Second},
 		"StandbyTaskMissingEventsDiscardDelay":                 {dynamicproperties.StandbyTaskMissingEventsDiscardDelay, time.Second},
-		"TimerDeletionOnWorkflowCloseMinTTL":                   {dynamicproperties.TimerDeletionOnWorkflowCloseMinTTL, time.Second},
+		"WorkflowTimerTaskCleanupMinTTL":                       {dynamicproperties.WorkflowTimerTaskCleanupMinTTL, time.Second},
 		"TaskProcessRPS":                                       {dynamicproperties.TaskProcessRPS, 30},
 		"TaskSchedulerType":                                    {dynamicproperties.TaskSchedulerType, 31},
 		"TaskSchedulerWorkerCount":                             {dynamicproperties.TaskSchedulerWorkerCount, 32},
